+++ src/main/java/com/santiago/priotti_api/Interfaces/Controller.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.santiago.priotti_web.Interfaces;
+package com.santiago.priotti_api.Interfaces;
 
 /**
  *
