+++ src/main/java/com/santiago/priotti_api/Interfaces/Dao.java
@@ -3,17 +3,19 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.santiago.priotti_web.Interfaces;
+package com.santiago.priotti_api.Interfaces;
+
+import java.sql.SQLException;
+import java.util.List;
 
 /**
  *
  * @author santiago
  * @param <T>
  */
-public interface Service {
-
-    public String getAll();
+public interface Dao<T> {
     
-    public String getAllAsDatatablesFormat();
-
+    
+    public List<T> get() throws SQLException;
+    
 }
