+++ src/main/java/com/santiago/priotti_api/Interfaces/Service.java
@@ -3,19 +3,17 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.santiago.priotti_web.Interfaces;
-
-import java.sql.SQLException;
-import java.util.List;
+package com.santiago.priotti_api.Interfaces;
 
 /**
  *
  * @author santiago
  * @param <T>
  */
-public interface Dao<T> {
+public interface Service {
+
+    public String getAll();
     
-    
-    public List<T> get() throws SQLException;
-    
+    public String getAllAsDatatablesFormat();
+
 }
