+++ src/main/java/com/santiago/priotti_api/StandardResponse/StatusResponse.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.santiago.priotti_web.StandardResponse;
+package com.santiago.priotti_api.StandardResponse;
 
 /**
  *
